Changed the type of the margin properties from long to decimal and saved the CompareCGPN property with the partner's config
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C31951

+++ CarrierPortal/Code/CDRComparison.BP.Arguments/Tasks/SettingTaskExecutionInformation.cs
@@ -11,9 +11,9 @@ namespace CDRComparison.BP.Arguments
     public class SettingTaskExecutionInformation : BPTaskExecutionInformation
     {
         public bool Decision { get; set; }
-        public long DurationMargin { get; set; }
+        public decimal DurationMargin { get; set; }
         public TimeUnitEnum DurationMarginTimeUnit { get; set; }
-        public long TimeMargin { get; set; }
+        public decimal TimeMargin { get; set; }
         public TimeUnitEnum TimeMarginTimeUnit { get; set; }
         public TimeSpan TimeOffset { get; set; }
         public bool CompareCGPN  { get; set; }
